refactor(crm): split general CRUD endpoints into invoke controllers
