+++ src/Crm/Transport/Controller/Api/V1/General/DeleteGeneralTaskController.php
@@ -1,0 +1,35 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Crm\Transport\Controller\Api\V1\General;
+
+use App\Crm\Domain\Entity\Task;
+use App\Role\Domain\Enum\Role;
+use Doctrine\ORM\EntityManagerInterface;
+use OpenApi\Attributes as OA;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Attribute\AsController;
+use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+
+#[AsController]
+#[OA\Tag(name: 'Crm')]
+#[IsGranted(Role::CRM_MANAGER->value)]
+final readonly class DeleteGeneralTaskController
+{
+    public function __construct(private EntityManagerInterface $entityManager)
+    {
+    }
+
+    #[Route('/v1/crm/general/tasks/{task}', methods: [Request::METHOD_DELETE])]
+    #[OA\Delete(summary: 'General - Delete Task', responses: [new OA\Response(response: 204, description: 'Task supprimée')])]
+    public function __invoke(Task $task): JsonResponse
+    {
+        $this->entityManager->remove($task);
+        $this->entityManager->flush();
+
+        return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
+    }
+}
