+++ src/Crm/Transport/Controller/Api/V1/General/ListGeneralBillingsController.php
@@ -1,0 +1,38 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Crm\Transport\Controller\Api\V1\General;
+
+use App\Crm\Domain\Entity\Billing;
+use App\Crm\Infrastructure\Repository\BillingRepository;
+use App\Role\Domain\Enum\Role;
+use OpenApi\Attributes as OA;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Attribute\AsController;
+use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+
+use function array_map;
+
+#[AsController]
+#[OA\Tag(name: 'Crm')]
+#[IsGranted(Role::CRM_MANAGER->value)]
+final readonly class ListGeneralBillingsController
+{
+    use GeneralCrudApiTrait;
+
+    public function __construct(private BillingRepository $billingRepository)
+    {
+    }
+
+    #[Route('/v1/crm/general/billings', methods: [Request::METHOD_GET])]
+    #[OA\Get(summary: 'General - List Billings', responses: [new OA\Response(response: 200, description: 'Liste des billings', content: new OA\JsonContent(example: ['items' => [['id' => 'uuid', 'label' => 'Abonnement avril', 'amount' => 99.9]]]))])]
+    public function __invoke(): JsonResponse
+    {
+        $items = array_map(fn (Billing $billing): array => $this->serializeBilling($billing), $this->billingRepository->findBy([], ['createdAt' => 'DESC']));
+
+        return new JsonResponse(['items' => $items]);
+    }
+}
